Add a tooltip that shows available combinations
When the user holds a key (for which different versions are available),
the script presents a tooltip showing the available combinations.
If AutoHotkey can retrieve the location of the text insertion point,
it will present it there, but this is often not the case,
so mostly it presents it near the cursor's location on screen.